requirements.txt
customtkinter==5.2.2
pillow==10.0.0
python-dotenv==1.1.0
requests==2.31.0
pyperclip==1.8.2